Add wiki_define() function
wiki_define() function displays plaintext extract(s) of the given
term(s) from the Wikipedia article(s).

It uses XML Wikipedia API, so it's very friendly to any internet
connection.

It is dependent on extension TextExtracts on a MediaWiki. However, it is
automatically enabled on English Wikipedia and comes with every
MediaWiki installation as of 1.34.0, so this should work properly. I am
leaving this note here only for case that in the future this
functionality would break.